fix every place to make new custody column function work

+++ beacon_chain/spec/datatypes/eip7594.nim
@@ -77,7 +77,7 @@ type
     row_index*: RowIndex
   
   CscBits* = BitArray[DATA_COLUMN_SIDECAR_SUBNET_COUNT]
-  CscCount* = uint8
+  CscCount* = uint64
 
   # https://github.com/ethereum/consensus-specs/blob/v1.5.0-alpha.4/specs/_features/eip7594/p2p-interface.md#metadata
   MetaData* = object
